# src/openapi-v3.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Guedder API",
    "description": "API do Guedder, a plataforma de venda de ingressos para eventos.",
    "version": "6.2.0"
  },
  "servers": [
    {
      "url": "https://api.dev.services.guedder.com",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v3/usuarios/perfil": {
      "get": {
        "tags": [
          "Usuario"
        ],
        "summary": "Retorna os dados do usuário autenticado",
        "description": "Retorna o perfil completo do usuário autenticado a partir do token JWT: identidade, contato, papel (`role`) e flags de permissão.  Legado: `GET /api/v1/usuarios/usuario_logado` ",
        "operationId": "getUsuarioLogado",
        "responses": {
          "200": {
            "description": "Dados do usuário autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioEntity"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioEntity"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v3/ingressos": {
      "get": {
        "tags": [
          "Ingresso"
        ],
        "summary": "Lista paginada dos ingressos do comprador autenticado",
        "description": "Legado: `GET /api/v2/ingresso/meus_ingressos/todos` Retorna a lista paginada dos ingressos do comprador. O agrupamento por status (ativos, encerrados, pendentes etc.) é responsabilidade do client. ",
        "operationId": "getMeusIngressos",
        "parameters": [
          {
            "name": "cicloDeVida",
            "in": "query",
            "description": "Filtro de ciclo de vida do ingresso: ATIVO (padrão), ENCERRADO ou TRANSFERENCIA_PENDENTE. ",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "ATIVO",
                "ENCERRADO",
                "TRANSFERENCIA_PENDENTE"
              ]
            }
          },
          {
            "name": "eventoId",
            "in": "query",
            "description": "Filtra os ingressos por evento.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Número da página (0-based).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Quantidade de itens por página.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Ordenação no formato campo,direção.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de ingressos do comprador",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedIngresso"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedIngresso"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v3/eventos/{eventoId}": {
      "get": {
        "tags": [
          "Evento"
        ],
        "summary": "Retorna os dados básicos de um evento por ID ou código",
        "description": "Retorna os dados básicos de um evento identificado por UUID ou por código alfanumérico. Endpoint público.  Legado: `GET /api/v1/evento/{id}` ",
        "operationId": "getEventoById",
        "parameters": [
          {
            "name": "eventoId",
            "in": "path",
            "description": "ID UUID ou código alfanumérico do evento.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados do evento",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DadosBasicosEventoDto"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/DadosBasicosEventoDto"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v3/eventos/{eventoId}/parametros-venda": {
      "get": {
        "tags": [
          "Checkout"
        ],
        "summary": "Retorna as formas de pagamento e regras de venda de um evento",
        "description": "Legado: `GET /api/v2/compra/evento/{eventoId}/parametros-venda/publico` Pode retornar 404 se o evento não tiver parâmetros configurados. ",
        "operationId": "getParametrosVenda",
        "parameters": [
          {
            "name": "eventoId",
            "in": "path",
            "description": "ID UUID ou código alfanumérico do evento.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Parâmetros de venda do evento",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParametrosVendaEvento"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ParametrosVendaEvento"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v3/eventos/{eventoId}/lotes": {
      "get": {
        "tags": [
          "Lote"
        ],
        "summary": "Lista os lotes disponíveis para compra em um evento",
        "description": "Lista os lotes disponíveis para compra em um evento (apenas os lotes públicos/visíveis). Endpoint público.  Legado: `GET /api/v1/lote/evento/{eventoId}/publico` ",
        "operationId": "listarLotesPublicos",
        "parameters": [
          {
            "name": "eventoId",
            "in": "path",
            "description": "ID UUID ou código alfanumérico do evento.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de lotes públicos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedLote"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedLote"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v3/eventos/{eventoId}/lotes/{loteId}": {
      "get": {
        "tags": [
          "Lote"
        ],
        "summary": "Retorna um lote pelo código ou ID do evento e do lote",
        "description": "Legado: `GET /api/v1/lote/{codigoOrEventoId}/lote/{codigoOrLoteId}` Aceita tanto IDs UUID quanto códigos alfanuméricos. ",
        "operationId": "getLotePorCodigoOuId",
        "parameters": [
          {
            "name": "eventoId",
            "in": "path",
            "description": "ID UUID ou código alfanumérico do evento (aceita UUID ou código).",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "loteId",
            "in": "path",
            "description": "ID UUID ou código alfanumérico do lote (aceita UUID ou código).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados do lote",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoteEntity"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/LoteEntity"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v3/eventos/{eventoId}/atracoes": {
      "get": {
        "tags": [
          "Atracoes"
        ],
        "summary": "Lista as atrações (line-up) de um evento",
        "description": "Lista as atrações (line-up) de um evento. Endpoint público; coleção bounded (retorna página única).  Legado: `GET /api/v1/evento/{eventoId}/atracoes/publico` ",
        "operationId": "listarAtracoesPorEvento",
        "parameters": [
          {
            "name": "eventoId",
            "in": "path",
            "description": "ID UUID ou código alfanumérico do evento.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de atrações",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedAtracao"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedAtracao"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v3/evento": {
      "get": {
        "tags": [
          "Evento"
        ],
        "summary": "Alias legado singular para listar eventos",
        "description": "Alias de compatibilidade para `GET /api/v3/eventos`. Aceita os mesmos filtros Sympla-compatible (`page`, `page_size`, `from`, `to`, `published`, `field_sort`, `sort`, `fields`, `timezone`). ",
        "operationId": "listarEventosLegacySingular",
        "responses": {
          "200": {
            "description": "Lista de eventos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedEvento"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedEvento"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit atingido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "deprecated": true
      }
    },
...
    "/api/v1/metrica/{eventoId}/ultimas-vendas": {
      "get": {
        "tags": [
          "metrica-evento-controller"
        ],
        "operationId": "getContagemIngressosVendidos",
        "parameters": [
          {
            "name": "eventoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RestObject"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetail": {
        "type": "object",
        "description": "Resposta de erro HTTP conforme RFC 7807. Gerado nativamente pelo Spring 6+ via `@ControllerAdvice` + `ProblemDetail`. ",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "example": "https://api.guedder.com/problems/not-found"
          },
          "title": {
            "type": "string",
            "example": "Not Found"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "example": 404
          },
          "detail": {
            "type": "string",
            "example": "Evento com id 'abc123' não encontrado."
          },
          "instance": {
            "type": "string",
            "format": "uri",
            "example": "https://api.guedder.com/api/v3/eventos/abc123"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiFieldError"
            }
          }
        }
      },
...
      "RestObject": {
        "type": "object",
        "additionalProperties": {},
        "properties": {
          "empty": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}